+++ lib/activerecord-import/adapters/sqlite3_adapter.rb
@@ -17,9 +17,8 @@ module ActiveRecord::Import::SQLite3Adapter
 
   # +sql+ can be a single string or an array. If it is an array all
   # elements that are in position >= 1 will be appended to the final SQL.
-  def insert_many(sql, values, *args) # :nodoc:
+  def insert_many( sql, values, _options = {}, *args ) # :nodoc:
     number_of_inserts = 0
-    ids = []
 
     base_sql, post_sql = if sql.is_a?( String )
       [sql, '']
@@ -30,22 +29,27 @@ module ActiveRecord::Import::SQLite3Adapter
     value_sets = ::ActiveRecord::Import::ValueSetsRecordsParser.parse(values,
       max_records: SQLITE_LIMIT_COMPOUND_SELECT)
 
-    value_sets.each do |value_set|
-      number_of_inserts += 1
-      sql2insert = base_sql + value_set.join( ',' ) + post_sql
-      first_insert_id = insert( sql2insert, *args )
-      last_insert_id = first_insert_id + value_set.size - 1
-      ids.concat((first_insert_id..last_insert_id).to_a)
+    transaction(requires_new: true) do
+      value_sets.each do |value_set|
+        number_of_inserts += 1
+        sql2insert = base_sql + value_set.join( ',' ) + post_sql
+        insert( sql2insert, *args )
+      end
     end
 
-    [number_of_inserts, ids]
+    [number_of_inserts, []]
+  end
+
+  def pre_sql_statements( options)
+    sql = []
+    # Options :recursive and :on_duplicate_key_ignore are mutually exclusive
+    if (options[:ignore] || options[:on_duplicate_key_ignore]) && !options[:recursive]
+      sql << "OR IGNORE"
+    end
+    sql + super
   end
 
   def next_value_for_sequence(sequence_name)
     %{nextval('#{sequence_name}')}
-  end
-
-  def support_setting_primary_key_of_imported_objects?
-    true
   end
 end