+++ test/support/factories.rb
@@ -32,6 +32,7 @@ FactoryGirl.define do
   end
 
   factory :rule do
+    sequence(:id) { |n| n }
     sequence(:condition_text) { |n| "q_#{n}_#{n}" }
   end
 
@@ -53,5 +54,10 @@ FactoryGirl.define do
   factory :book do
     title 'Tortilla Flat'
     author_name 'John Steinbeck'
+  end
+
+  factory :car do
+    sequence(:Name) { |n| n }
+    sequence(:Features) { |n| "Feature #{n}" }
   end
 end